Parallel IO and Art-Net can be switched between. Still just 1 bus.
